Remove support for prior to Go 1.18 in CLI
