--- cmd/ensure/internal/ifacereader/ifacereader_go118.go
@@ -1,32 +1,0 @@
-//go:build go1.18
-// +build go1.18
-
-package ifacereader
-
-import (
-	"go/types"
-)
-
-func (r *internalPackageReader) parseTypeParams(namedType *types.Named) []*TypeParam {
-	typeParamList := namedType.TypeParams()
-	typeParamCount := typeParamList.Len()
-
-	if typeParamCount == 0 {
-		return nil
-	}
-
-	typeParams := make([]*TypeParam, 0, typeParamCount)
-	//nolint:intrange // Requires Go 1.20+
-	for i := 0; i < typeParamCount; i++ {
-		typeParam := typeParamList.At(i)
-
-		typeParams = append(typeParams, &TypeParam{
-			Name: typeParam.Obj().Name(),
-			Type: types.TypeString(typeParam.Constraint(), func(p *types.Package) string {
-				return r.pkgNameGen.GeneratePackageName(r.pkg, p)
-			}),
-		})
-	}
-
-	return typeParams
-}
