--- cmd/ensure/internal/ifacereader/ifacereader_pre_go118_test.go
@@ -1,8 +1,0 @@
-//go:build !go1.18
-// +build !go1.18
-
-package ifacereader_test
-
-func buildGenericTests() []entry {
-	return nil
-}
